:hammer: move mapToColor to new file
